Provas/Modulo10/prova2/src/blogpost/requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
Jinja2==3.1.4
markdown-it-py==3.0.0
orjson==3.10.4
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
SQLAlchemy==2.0.30
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
